exception-handlers/src/main/java/fi/tuomas/chassis/exceptionhandlers/advice/GeneralControllerAdvice.java
package fi.tuomas.chassis.exceptionhandlers.advice;

import org.springframework.http.HttpStatus;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestControllerAdvice;

@RestControllerAdvice
public class GeneralControllerAdvice {

    @ExceptionHandler(MethodArgumentNotValidException.class)
    @ResponseStatus(HttpStatus.BAD_REQUEST)
    public String handleException(MethodArgumentNotValidException e) {
        return String.format("Number of errors: %d", e.getErrorCount());
    }
}
